src/requirements.txt
boto3==1.26.76
email-validator==1.3.1
Jinja2==3.1.2
pydantic==1.10.5
python-dateutil==2.8.2